requirements.txt
Transformers
Transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
pyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box == 6.0.2
ensure == 1.0.2
fastapi == 0.78.0
unicorn == 2.0.1.post1
jinja2 == 3.1.2
-e .